fix(LLM): reduce app rerenders with redux store updates
