+++ .changeset/smooth-dogs-battle.md
@@ -1,0 +1,5 @@
+---
+"live-mobile": minor
+---
+
+move useDBSaveEffect usage away from LLM root App
