+++ apps/ledger-live-mobile/src/components/DBSave.ts
@@ -1,9 +1,33 @@
-import { useRef, useEffect, useMemo } from "react";
+import { useRef, useEffect, useMemo, useCallback } from "react";
 import identity from "lodash/identity";
 import throttleFn from "lodash/throttle";
-import { useSelector } from "react-redux";
 import { Maybe } from "../types/helpers";
-import { State } from "../reducers/types";
+import { useCountervaluesExport } from "@ledgerhq/live-countervalues-react";
+import { pairId } from "@ledgerhq/live-countervalues/helpers";
+import isEqual from "lodash/isEqual";
+import { postOnboardingSelector } from "@ledgerhq/live-common/postOnboarding/reducer";
+import { useSelector } from "react-redux";
+import {
+  saveAccounts,
+  saveBle,
+  saveSettings,
+  saveCountervalues,
+  savePostOnboardingState,
+  saveMarketState,
+  saveTrustchainState,
+  saveWalletExportState,
+  saveLargeMoverState,
+} from "~/db";
+import { exportSelector as settingsExportSelector } from "~/reducers/settings";
+import { exportSelector as accountsExportSelector } from "~/reducers/accounts";
+import { exportSelector as bleSelector } from "~/reducers/ble";
+import type { State } from "~/reducers/types";
+import { useTrackingPairs } from "~/actions/general";
+import { trustchainStoreSelector } from "@ledgerhq/ledger-key-ring-protocol/store";
+import { walletSelector } from "~/reducers/wallet";
+import { exportWalletState, walletStateExportShouldDiffer } from "@ledgerhq/live-wallet/store";
+import { exportMarketSelector } from "~/reducers/market";
+import { exportLargeMoverSelector } from "~/reducers/largeMover";
 
 type MaybeState = Maybe<State>;
 
@@ -13,12 +37,8 @@ type Props<Data, Stats> = {
   getChangesStats: (next: State, prev: State) => Stats;
   save: (data: Data, changedStats: Stats) => Promise<void>;
 };
-export default function useDBSaveEffect<D, S>({
-  lense,
-  throttle = 500,
-  save,
-  getChangesStats,
-}: Props<D, S>) {
+
+function useDBSaveEffect<D, S>({ lense, throttle = 500, save, getChangesStats }: Props<D, S>) {
   const state: MaybeState = useSelector(identity);
   const lastSavedState = useRef(state);
   // we keep an updated version of current props in "latestProps" ref
@@ -73,3 +93,108 @@ function useFlushMechanism({ flush, cancel }: { flush: () => void; cancel: () =>
     };
   }, [flush]);
 }
+
+function walletExportSelector(state: State) {
+  return exportWalletState(walletSelector(state));
+}
+
+export const ConfigureDBSaveEffects = () => {
+  const getSettingsChanged = useCallback((a: State, b: State) => a.settings !== b.settings, []);
+  const getAccountsChanged = useCallback(
+    (
+      oldState: State,
+      newState: State,
+    ):
+      | {
+          changed: string[];
+        }
+      | null
+      | undefined => {
+      if (oldState.accounts !== newState.accounts) {
+        return {
+          changed: newState.accounts.active
+            .filter(a => {
+              const old = oldState.accounts.active.find(b => a.id === b.id);
+              return !old || old !== a;
+            })
+            .map(a => a.id),
+        };
+      }
+      return null;
+    },
+    [],
+  );
+
+  const getPostOnboardingStateChanged = useCallback(
+    (a: State, b: State) => !isEqual(a.postOnboarding, b.postOnboarding),
+    [],
+  );
+
+  const rawState = useCountervaluesExport();
+  const trackingPairs = useTrackingPairs();
+  const pairIds = useMemo(() => trackingPairs.map(p => pairId(p)), [trackingPairs]);
+
+  useDBSaveEffect({
+    save: saveCountervalues,
+    throttle: 2000,
+    getChangesStats: () => ({
+      changed: !!Object.keys(rawState.status).length,
+      pairIds,
+    }),
+    lense: () => rawState,
+  });
+  useDBSaveEffect({
+    save: saveSettings,
+    throttle: 400,
+    getChangesStats: getSettingsChanged,
+    lense: settingsExportSelector,
+  });
+  useDBSaveEffect({
+    save: saveAccounts,
+    throttle: 500,
+    getChangesStats: getAccountsChanged,
+    lense: accountsExportSelector,
+  });
+  useDBSaveEffect({
+    save: saveBle,
+    throttle: 500,
+    getChangesStats: (a, b) => a.ble !== b.ble,
+    lense: bleSelector,
+  });
+  useDBSaveEffect({
+    save: savePostOnboardingState,
+    throttle: 500,
+    getChangesStats: getPostOnboardingStateChanged,
+    lense: postOnboardingSelector,
+  });
+
+  useDBSaveEffect({
+    save: saveMarketState,
+    throttle: 500,
+    getChangesStats: (a, b) => a.market !== b.market,
+    lense: exportMarketSelector,
+  });
+
+  useDBSaveEffect({
+    save: saveTrustchainState,
+    throttle: 500,
+    getChangesStats: (a, b) => a.trustchain !== b.trustchain,
+    lense: trustchainStoreSelector,
+  });
+
+  useDBSaveEffect({
+    save: saveWalletExportState,
+    throttle: 500,
+    getChangesStats: (a, b) => walletStateExportShouldDiffer(a.wallet, b.wallet),
+    lense: walletExportSelector,
+  });
+
+  useDBSaveEffect({
+    save: saveLargeMoverState,
+    throttle: 500,
+    getChangesStats: (a, b) => a.largeMover !== b.largeMover,
+    lense: exportLargeMoverSelector,
+  });
+
+  return null;
+};
