+++ apps/ledger-live-mobile/src/index.tsx
@@ -11,11 +11,7 @@ import Transport from "@ledgerhq/hw-transport";
 import { NotEnoughBalance } from "@ledgerhq/errors";
 import { log } from "@ledgerhq/logs";
 import { checkLibs } from "@ledgerhq/live-common/sanityChecks";
-import { useCountervaluesExport } from "@ledgerhq/live-countervalues-react";
-import { pairId } from "@ledgerhq/live-countervalues/helpers";
 import "./config/configInit";
-import isEqual from "lodash/isEqual";
-import { postOnboardingSelector } from "@ledgerhq/live-common/postOnboarding/reducer";
 import Config from "react-native-config";
 import { LogLevel, PerformanceProfiler, RenderPassReport } from "@shopify/react-native-performance";
 import useEnv from "@ledgerhq/live-common/hooks/useEnv";
@@ -23,26 +19,13 @@ import { useDispatch, useSelector } from "react-redux";
 import { init } from "../e2e/bridge/client";
 import logger from "./logger";
 import {
-  saveAccounts,
-  saveBle,
-  saveSettings,
-  saveCountervalues,
-  savePostOnboardingState,
-  saveMarketState,
-  saveTrustchainState,
-  saveWalletExportState,
-  saveLargeMoverState,
-} from "./db";
-import {
-  exportSelector as settingsExportSelector,
   osThemeSelector,
   hasSeenAnalyticsOptInPromptSelector,
   hasCompletedOnboardingSelector,
   trackingEnabledSelector,
   reportErrorsEnabledSelector,
 } from "~/reducers/settings";
-import { accountsSelector, exportSelector as accountsExportSelector } from "~/reducers/accounts";
-import { exportSelector as bleSelector } from "~/reducers/ble";
+import { accountsSelector } from "~/reducers/accounts";
 import LocaleProvider, { i18n } from "~/context/Locale";
 import RebootProvider from "~/context/Reboot";
 import AuthPass from "~/context/AuthPass";
@@ -52,15 +35,12 @@ import LoadingApp from "~/components/LoadingApp";
 import StyledStatusBar from "~/components/StyledStatusBar";
 import AnalyticsConsole from "~/components/AnalyticsConsole";
 import DebugTheme from "~/components/DebugTheme";
-import useDBSaveEffect from "~/components/DBSave";
 import SyncNewAccounts from "~/bridge/SyncNewAccounts";
 import SegmentSetup from "~/analytics/SegmentSetup";
 import HookSentry from "~/components/HookSentry";
 import HookNotifications from "~/notifications/HookNotifications";
 import RootNavigator from "~/components/RootNavigator";
 import SetEnvsFromSettings from "~/components/SetEnvsFromSettings";
-import type { State } from "~/reducers/types";
-import { useTrackingPairs } from "~/actions/general";
 import ExperimentalHeader from "~/screens/Settings/Experimental/ExperimentalHeader";
 import Modals from "~/screens/Modals";
 import NavBarColorHandler from "~/components/NavBarColorHandler";
@@ -88,17 +68,12 @@ import { useSettings } from "~/hooks";
 import AppProviders from "./AppProviders";
 import { useAutoDismissPostOnboardingEntryPoint } from "@ledgerhq/live-common/postOnboarding/hooks/index";
 import QueuedDrawersContextProvider from "LLM/components/QueuedDrawer/QueuedDrawersContextProvider";
-import { exportMarketSelector } from "./reducers/market";
-import { trustchainStoreSelector } from "@ledgerhq/ledger-key-ring-protocol/store";
-import { walletSelector } from "~/reducers/wallet";
-import { exportWalletState, walletStateExportShouldDiffer } from "@ledgerhq/live-wallet/store";
 import { registerTransports } from "~/services/registerTransports";
 import { useDeviceManagementKitEnabled } from "@ledgerhq/live-dmk-mobile";
 import { StoragePerformanceOverlay } from "./newArch/storage/screens/PerformanceMonitor";
 import { useDeviceManagementKit } from "@ledgerhq/live-dmk-mobile";
 import AppVersionBlocker from "LLM/features/AppBlockers/components/AppVersionBlocker";
 import AppGeoBlocker from "LLM/features/AppBlockers/components/AppGeoBlocker";
-import { exportLargeMoverSelector } from "./reducers/largeMover";
 import {
   TrackingConsent,
   DatadogProvider,
@@ -113,6 +88,7 @@ import getOrCreateUser from "./user";
 import { FIRST_PARTY_MAIN_HOST_DOMAIN } from "./utils/constants";
 import { BottomSheetModalProvider } from "@gorhom/bottom-sheet";
 import useNativeStartupInfo from "./hooks/useNativeStartupInfo";
+import { ConfigureDBSaveEffects } from "./components/DBSave";
 
 if (Config.DISABLE_YELLOW_BOX) {
   LogBox.ignoreAllLogs();
@@ -134,10 +110,6 @@ const styles = StyleSheet.create({
     flex: 1,
   },
 });
-
-function walletExportSelector(state: State) {
-  return exportWalletState(walletSelector(state));
-}
 
 function App() {
   const accounts = useSelector(accountsSelector);
@@ -230,104 +202,9 @@ function App() {
   useListenToHidDevices();
   useAutoDismissPostOnboardingEntryPoint();
 
-  const getSettingsChanged = useCallback((a: State, b: State) => a.settings !== b.settings, []);
-  const getAccountsChanged = useCallback(
-    (
-      oldState: State,
-      newState: State,
-    ):
-      | {
-          changed: string[];
-        }
-      | null
-      | undefined => {
-      if (oldState.accounts !== newState.accounts) {
-        return {
-          changed: newState.accounts.active
-            .filter(a => {
-              const old = oldState.accounts.active.find(b => a.id === b.id);
-              return !old || old !== a;
-            })
-            .map(a => a.id),
-        };
-      }
-      return null;
-    },
-    [],
-  );
-
-  const getPostOnboardingStateChanged = useCallback(
-    (a: State, b: State) => !isEqual(a.postOnboarding, b.postOnboarding),
-    [],
-  );
-
-  const rawState = useCountervaluesExport();
-  const trackingPairs = useTrackingPairs();
-  const pairIds = useMemo(() => trackingPairs.map(p => pairId(p)), [trackingPairs]);
-  useDBSaveEffect({
-    save: saveCountervalues,
-    throttle: 2000,
-    getChangesStats: () => ({
-      changed: !!Object.keys(rawState.status).length,
-      pairIds,
-    }),
-    lense: () => rawState,
-  });
-  useDBSaveEffect({
-    save: saveSettings,
-    throttle: 400,
-    getChangesStats: getSettingsChanged,
-    lense: settingsExportSelector,
-  });
-  useDBSaveEffect({
-    save: saveAccounts,
-    throttle: 500,
-    getChangesStats: getAccountsChanged,
-    lense: accountsExportSelector,
-  });
-  useDBSaveEffect({
-    save: saveBle,
-    throttle: 500,
-    getChangesStats: (a, b) => a.ble !== b.ble,
-    lense: bleSelector,
-  });
-  useDBSaveEffect({
-    save: savePostOnboardingState,
-    throttle: 500,
-    getChangesStats: getPostOnboardingStateChanged,
-    lense: postOnboardingSelector,
-  });
-
-  useDBSaveEffect({
-    save: saveMarketState,
-    throttle: 500,
-    getChangesStats: (a, b) => a.market !== b.market,
-    lense: exportMarketSelector,
-  });
-
-  useDBSaveEffect({
-    save: saveTrustchainState,
-    throttle: 500,
-    getChangesStats: (a, b) => a.trustchain !== b.trustchain,
-    lense: trustchainStoreSelector,
-  });
-
-  useDBSaveEffect({
-    save: saveWalletExportState,
-    throttle: 500,
-    getChangesStats: (a, b) => walletStateExportShouldDiffer(a.wallet, b.wallet),
-    lense: walletExportSelector,
-  });
-
-  useDBSaveEffect({
-    save: saveLargeMoverState,
-    throttle: 500,
-    getChangesStats: (a, b) => a.largeMover !== b.largeMover,
-    lense: exportLargeMoverSelector,
-  });
-
   return (
     <>
+      <ConfigureDBSaveEffects />
       <SyncNewAccounts priority={5} />
       <TransactionsAlerts />
       <ExperimentalHeader />
